+++ libs/dh/reports/feature-settlement-reports/src/components/details/details.component.html
