+++ libs/dh/reports/feature-settlement-reports/src/components/details/duration.component.ts
