+++ libs/dh/reports/feature-settlement-reports/src/components/overview/overview.component.html
@@ -84,7 +84,7 @@ limitations under the License.
   </watt-table>
 </vater-flex>
 
-<dh-report-details
+<dh-details
   #reportDetails
   (closed)="activeRow.set(undefined)"
   (download)="downloadReport($event, activeRow()!)"
