+++ libs/dh/reports/feature-settlement-reports/src/components/request-report/request-as-modal.component.html
@@ -14,7 +14,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 -->
-
 <watt-modal
   #modal
   [title]="t('title')"
